Harden triangulation and fix robustness issues in Tydra mesh pipeline
Triangulation improvements:
- Newell normal: compute entirely in double to prevent float cancellation
- Degenerate polygons: skip with warning instead of aborting entire mesh
- Earcut temporaries: reuse polyline/polygon_2d across faces (no per-face alloc)
- triangulatedToOrigFaceVertexIndexMap: narrow size_t to uint32_t (halve memory)
- Quad split: choose shorter diagonal for better triangle quality

Normal/tangent robustness:
- GeometricNormal: guard against degenerate triangles (zero-area → zero normal)
- ComputeNormals: skip zero-area faces, guard final normalization
- BlendShape: fix off-by-one bounds check (> → >=), eliminate redundant copy
- Guard std::max_element against empty containers (UB prevention)

Memory reduction:
- Remove redundant memset on zero-initialized tn/bn vectors
- Eliminate duplicate facevarying normal expansion in tangent computation

+++ src/tydra/render-data.hh
@@ -1160,7 +1160,7 @@ struct RenderMesh {
   std::vector<uint32_t> triangulatedFaceVertexIndices;
   std::vector<uint32_t> triangulatedFaceVertexCounts;
 
-  std::vector<size_t>
+  std::vector<uint32_t>
       triangulatedToOrigFaceVertexIndexMap;  // used for rearrange facevertex
                                              // attrib
   std::vector<uint32_t>
